+++ database/migrations/2023_07_03_115003_create_cardiovascular_risks_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('cardiovascular_risks', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('client_id');//foreign. Is necessary the unsigned to match with the other table
+            $table->unsignedBigInteger('client_id');//foreign. Is necessary the unsigned to match with the other table
             $table->enum('risk', ['low', 'medium', 'high']);
             $table->softDeletes();
             $table->timestamps();
