+++ database/migrations/2023_07_03_115843_create_nutrition_and_health_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('nutrition_and_health', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('client_id');//foreign. Is necessary the unsigned to match with the other table
+            $table->unsignedBigInteger('client_id');//foreign. Is necessary the unsigned to match with the other table
             $table->json('recommendations');
             $table->timestamps();
         });
